+++ src/main/java/io/supertokens/webserver/InputParser.java
@@ -86,6 +86,21 @@ public class InputParser {
         }
     }
 
+    public static Long getLongQueryParamOrThrowError(HttpServletRequest request, String fieldName, boolean nullable)
+            throws ServletException {
+        String key = getQueryParamOrThrowError(request, fieldName, nullable);
+        if (key == null && nullable) {
+            return null;
+        }
+        try {
+            assert key != null;
+            return Long.parseLong(key);
+        } catch (Exception e) {
+            throw new ServletException(new WebserverAPI.BadRequestException(
+                    "Field name '" + fieldName + "' must be a long in the GET request"));
+        }
+    }
+
     public static JsonObject parseJsonObjectOrThrowError(JsonObject element, String fieldName, boolean nullable)
             throws ServletException {
         try {
@@ -197,4 +212,28 @@ public class InputParser {
         }
 
     }
+
+
+    public static Integer parseIntOrThrowError(JsonObject element, String fieldName, boolean nullable)
+            throws ServletException {
+        try {
+            if (nullable && element.get(fieldName) == null) {
+                return null;
+
+            }
+            String stringified = element.toString();
+            if (!stringified.contains("\"")) {
+                throw new Exception();
+
+            }
+            return element.get(fieldName).getAsInt();
+
+        } catch (Exception e) {
+            throw new ServletException(
+                    new WebserverAPI.BadRequestException("Field name '" + fieldName + "' is invalid in JSON input"));
+
+        }
+
+    }
+
 }