+++ src/test/java/io/supertokens/test/totp/api/RemoveTotpDeviceAPITest.java
@@ -1,0 +1,188 @@
+package io.supertokens.test.totp.api;
+
+import com.google.gson.JsonObject;
+import io.supertokens.ProcessState;
+import io.supertokens.featureflag.EE_FEATURES;
+import io.supertokens.featureflag.FeatureFlag;
+import io.supertokens.featureflag.FeatureFlagTestContent;
+import io.supertokens.test.httpRequest.HttpResponseException;
+import io.supertokens.pluginInterface.STORAGE_TYPE;
+import io.supertokens.storageLayer.StorageLayer;
+import io.supertokens.test.TestingProcessManager;
+import io.supertokens.test.Utils;
+import io.supertokens.test.httpRequest.HttpRequestForTesting;
+import io.supertokens.test.totp.TotpLicenseTest;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TestRule;
+
+import static org.junit.Assert.*;
+
+public class RemoveTotpDeviceAPITest {
+
+    @Rule
+    public TestRule watchman = Utils.getOnFailure();
+
+    @AfterClass
+    public static void afterTesting() {
+        Utils.afterTesting();
+    }
+
+    @Before
+    public void beforeEach() {
+        Utils.reset();
+    }
+
+    private Exception removeDeviceRequest(TestingProcessManager.TestingProcess process, JsonObject body) {
+        return assertThrows(
+                io.supertokens.test.httpRequest.HttpResponseException.class,
+                () -> HttpRequestForTesting.sendJsonPOSTRequest(
+                        process.getProcess(),
+                        "",
+                        "http://localhost:3567/recipe/totp/device/remove",
+                        body,
+                        1000,
+                        1000,
+                        null,
+                        Utils.getCdiVersionLatestForTests(),
+                        "totp"));
+    }
+
+    private void checkFieldMissingErrorResponse(Exception ex, String fieldName) {
+        assert ex instanceof HttpResponseException;
+        HttpResponseException e = (HttpResponseException) ex;
+        assert e.statusCode == 400;
+        assertTrue(e.getMessage().contains(
+                "Http error. Status Code: 400. Message: Field name '" + fieldName + "' is invalid in JSON input"));
+    }
+
+    private void checkResponseErrorContains(Exception ex, String msg) {
+        assert ex instanceof HttpResponseException;
+        HttpResponseException e = (HttpResponseException) ex;
+        assert e.statusCode == 400;
+        assertTrue(e.getMessage().contains(msg));
+    }
+
+    @Test
+    public void testApi() throws Exception {
+        String[] args = { "../" };
+
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
+            return;
+        }
+
+        FeatureFlagTestContent.getInstance(process.main).setKeyValue(FeatureFlagTestContent.ENABLED_FEATURES, new EE_FEATURES[] { EE_FEATURES.TOTP });
+
+        // Setup user and devices:
+        JsonObject createDeviceReq = new JsonObject();
+        createDeviceReq.addProperty("userId", "user-id");
+        createDeviceReq.addProperty("deviceName", "d1");
+        createDeviceReq.addProperty("period", 30);
+        createDeviceReq.addProperty("skew", 0);
+
+        JsonObject createDeviceRes = HttpRequestForTesting.sendJsonPOSTRequest(
+                process.getProcess(),
+                "",
+                "http://localhost:3567/recipe/totp/device",
+                createDeviceReq,
+                1000,
+                1000,
+                null,
+                Utils.getCdiVersionLatestForTests(),
+                "totp");
+        assertEquals(createDeviceRes.get("status").getAsString(), "OK");
+
+        // create another device d2:
+        createDeviceReq.addProperty("deviceName", "d2");
+        JsonObject createDeviceRes2 = HttpRequestForTesting.sendJsonPOSTRequest(
+                process.getProcess(),
+                "",
+                "http://localhost:3567/recipe/totp/device",
+                createDeviceReq,
+                1000,
+                1000,
+                null,
+                Utils.getCdiVersionLatestForTests(),
+                "totp");
+        assertEquals(createDeviceRes2.get("status").getAsString(), "OK");
+
+        // Start the actual tests for remove device API:
+
+        JsonObject body = new JsonObject();
+
+        // Missing userId/deviceName
+        {
+            Exception e = removeDeviceRequest(process, body);
+            checkFieldMissingErrorResponse(e, "userId");
+
+            body.addProperty("userId", "");
+            e = removeDeviceRequest(process, body);
+            checkFieldMissingErrorResponse(e, "deviceName");
+
+        }
+
+        // Invalid userId/deviceName
+        {
+            body.addProperty("deviceName", "");
+            Exception e = removeDeviceRequest(process, body);
+            checkResponseErrorContains(e, "userId cannot be empty"); // Note that this is not a field missing error
+
+            body.addProperty("userId", "user-id");
+            e = removeDeviceRequest(process, body);
+            checkResponseErrorContains(e, "deviceName cannot be empty");
+
+            body.addProperty("deviceName", "d1");
+
+            // should pass now:
+            JsonObject res = HttpRequestForTesting.sendJsonPOSTRequest(
+                    process.getProcess(),
+                    "",
+                    "http://localhost:3567/recipe/totp/device/remove",
+                    body,
+                    1000,
+                    1000,
+                    null,
+                    Utils.getCdiVersionLatestForTests(),
+                    "totp");
+            assert res.get("status").getAsString().equals("OK");
+            assert res.get("didDeviceExist").getAsBoolean() == true;
+
+            // try again with same device (still pass but didDeviceExist should be false)
+            JsonObject res2 = HttpRequestForTesting.sendJsonPOSTRequest(
+                    process.getProcess(),
+                    "",
+                    "http://localhost:3567/recipe/totp/device/remove",
+                    body,
+                    1000,
+                    1000,
+                    null,
+                    Utils.getCdiVersionLatestForTests(),
+                    "totp");
+            assert res2.get("status").getAsString().equals("OK");
+            assert res2.get("didDeviceExist").getAsBoolean() == false;
+
+            // try deleting device for a non-existent user
+            body.addProperty("userId", "non-existent-user");
+            JsonObject res3 = HttpRequestForTesting.sendJsonPOSTRequest(
+                    process.getProcess(),
+                    "",
+                    "http://localhost:3567/recipe/totp/device/remove",
+                    body,
+                    1000,
+                    1000,
+                    null,
+                    Utils.getCdiVersionLatestForTests(),
+                    "totp");
+            assert res3.get("status").getAsString().equals("TOTP_NOT_ENABLED_ERROR");
+        }
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+}